+++ pharus/dynamic_api_gen.py
@@ -27,10 +27,10 @@ def {method_name}(jwt_payload: dict) -> dict:
             djconn = _DJConnector._set_datajoint_config(jwt_payload)
             vm_list = [dj.VirtualModule(s, s, connection=djconn)
                        for s in inspect.getfullargspec(dj_query).args]
-            query, fetch_args = dj_query(*vm_list)
-            query = query & restriction()
+            djdict = dj_query(*vm_list)
+            djdict['query'] = djdict['query'] & restriction()
             record_header, table_tuples, total_count = _DJConnector._fetch_records(
-                query=query,
+                query=djdict['query'], fetch_args=djdict['fetch_args'],
                 **{{k: (int(v) if k in ('limit', 'page')
                    else (v.split(',') if k == 'order'
                    else loads(b64decode(v.encode('utf-8')).decode('utf-8'))))
@@ -52,14 +52,36 @@ def {method_name}_attributes(jwt_payload: dict) -> dict:
             djconn = _DJConnector._set_datajoint_config(jwt_payload)
             vm_list = [dj.VirtualModule(s, s, connection=djconn)
                        for s in inspect.getfullargspec(dj_query).args]
-            query, fetch_args = dj_query(*vm_list)
-            attributes_meta = _DJConnector._get_attributes(query)
+            djdict = dj_query(*vm_list)
+            attributes_meta = _DJConnector._get_attributes(djdict['query'])
 
             return dict(attributeHeaders=attributes_meta['attribute_headers'],
                         attributes=attributes_meta['attributes'])
         except Exception as e:
             return str(e), 500
 """
+
+    plot_route_template = '''
+
+@app.route('{route}', methods=['GET'])
+@protected_route
+def {method_name}(jwt_payload: dict) -> dict:
+
+{query}
+{restriction}
+    if request.method in {{'GET'}}:
+        try:
+            djconn = _DJConnector._set_datajoint_config(jwt_payload)
+            vm_list = [dj.VirtualModule(s, s, connection=djconn)
+                       for s in inspect.getfullargspec(dj_query).args]
+            djdict = dj_query(*vm_list)
+            djdict['query'] = djdict['query'] & restriction()
+            record_header, table_tuples, total_count = _DJConnector._fetch_records(
+                fetch_args=djdict['fetch_args'], query=djdict['query'], fetch_blobs=True)
+            return dict(table_tuples[0][0])
+        except Exception as e:
+            return str(e), 500
+'''
 
     pharus_root = f"{pkg_resources.get_distribution('pharus').module_path}/pharus"
     api_path = f'{pharus_root}/dynamic_api.py'
@@ -76,6 +98,11 @@ def {method_name}_attributes(jwt_payload: dict) -> dict:
                 for comp in grid['components'].values():
                     if comp['type'] == 'table':
                         f.write(route_template.format(route=comp['route'],
+                                method_name=comp['route'].replace('/', ''),
+                                query=indent(comp['dj_query'], '    '),
+                                restriction=indent(comp['restriction'], '    ')))
+                    if comp['type'] == 'plot:plotly:stored_json':
+                        f.write(plot_route_template.format(route=comp['route'],
                                 method_name=comp['route'].replace('/', ''),
                                 query=indent(comp['dj_query'], '    '),
                                 restriction=indent(comp['restriction'], '    ')))