+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Changelog
 
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
+## [0.2.3] - 2021-11-18
+### Added
+- Support for plot component PR #155
+- Fetch argument specification in `dj_query` PR #155
 
 ## [0.2.2] - 2021-11-10
 ### Fixed
-- Optimize dynamic api virtual modules.
+- Optimize dynamic api virtual modules. PR #113
 
 ## [0.2.1] - 2021-11-08
 ### Fixed
-- Error with retrieving the module's installation root path.
+- Error with retrieving the module's installation root path. PR #112
 
 ## [0.2.0] - 2021-11-02
 ### Added
@@ -84,6 +88,7 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 - Support for DataJoint attribute types: `varchar`, `int`, `float`, `datetime`, `date`, `time`, `decimal`, `uuid`.
 - Check dependency utility to determine child table references.
 
+[0.2.3]: https://github.com/datajoint/pharus/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/datajoint/pharus/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/datajoint/pharus/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/datajoint/pharus/compare/0.1.0...0.2.0